requirements.txt
# Dependências do projeto
ultralytics>=8.0.20         
opencv-python>=4.7.0
numpy>=1.23.0
pandas>=1.5.0
Pillow>=9.3.0
openpyxl>=3.1.0
